% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_create_local_database.R
\name{download_create_local_database}
\alias{download_create_local_database}
\title{download_create_local_database Title}
\usage{
download_create_local_database(con)
}
\value{
1
}
\description{
This function downloads a database file from a remote location and creates a connection to it.
}
